Problem modification combined with clarke & wright, worst fit and local search implementations.
Model file contains all the necessary classes with their functions used to initialize and execute the problem model. Uniform_routes file contains an implementation of the worst fit bin packing problem adjusted to the problem structure, as well as the k-n swaps local search operator combined with the possibility to also implement the multi restart method.